moved files into public and src folders
