--- components/grid/tabl.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import Tabl_row from './tabl_row.js';
-
-export default ({ name }) => 
-
-<table id="grid">
-  <tbody>
-    <Tabl_row />
-    <Tabl_row />
-    <Tabl_row />
-    <Tabl_row />
-    <Tabl_row />
-  </tbody>
-</table>;
