--- index.html
@@ -1,3 +1,0 @@
-<link href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
-
-<div id="root"></div>
