+++ src/components/grid/tabl_row.js
@@ -1,0 +1,26 @@
+import React from 'react';
+import Beat_btn from './beat-btn.js';
+
+export default ({ name }) => 
+
+<tr className="table-row">
+  <td className="table-col">
+    <Beat_btn />
+  </td>
+
+  <td className="table-col">
+    <Beat_btn />
+  </td>
+
+  <td className="table-col">
+    <Beat_btn />
+  </td>
+
+  <td className="table-col">
+    <Beat_btn />
+  </td>
+
+  <td className="table-col">
+    <Beat_btn />
+  </td>
+</tr>;
